added adjust on login and css pages
