+++ front/src/Pages/private/StatusCode/index.css
@@ -2,7 +2,7 @@
     background: url(../../../Assets/bgstatus.gif);
     background-attachment: fixed;
 	background-position: center center;
-    background-size: contain;
+    background-size: cover;
     min-width: 98vw;
     min-height: 99.5vh;
     display: flex;
